Edit add-remove.js,Rename remove-edit.js,Update markUp.js
